Correcoes CategoriaService e Pagamento

+++ src/main/java/com/franciscoaguiar/cursomc/services/ClienteService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
@@ -17,40 +16,41 @@ import com.franciscoaguiar.cursomc.services.exceptions.ObjectNotFoundException;
 
 @Service
 public class ClienteService {
-	
+
 	@Autowired
 	private ClienteRepository repository;
-	
+
 	public Cliente find(Integer id) {
 		Optional<Cliente> Cliente = repository.findById(id);
-		return Cliente.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado! Id: " + id + ", Tipo: " + Cliente.class.getName()));
+		return Cliente.orElseThrow(() -> new ObjectNotFoundException(
+				"Objeto não encontrado! Id: " + id + ", Tipo: " + Cliente.class.getName()));
 	}
-	
+
 	public Cliente update(Cliente obj) {
 		Cliente cli = repository.findById(obj.getId()).get();
-		updateData(obj,cli);
+		updateData(obj, cli);
 		return repository.save(obj);
 	}
-	
 
 	public void delete(Integer id) {
-		try {
-			repository.deleteById(id);			
+		Cliente cli = repository.findById(id).get();
+		if (cli.getPedidos().isEmpty()) {
+			repository.deleteById(id);
+		} else {
+			throw new DataIntegrityException("Não é possível excluir um cliente, pois há pedidos relacionadas");
 		}
-		catch (DataIntegrityViolationException e) {
-			throw new DataIntegrityException("Não é possível excluir um cliente, pois há entidades relacionadas");
-		}
+
 	}
 
 	public List<Cliente> findAll() {
 		return repository.findAll();
 	}
-	
-	public Page<Cliente> findPage(Integer page, Integer linesPerPage, String orderBy, String direction){
+
+	public Page<Cliente> findPage(Integer page, Integer linesPerPage, String orderBy, String direction) {
 		PageRequest pageRequest = PageRequest.of(page, linesPerPage, Direction.valueOf(direction), orderBy);
 		return repository.findAll(pageRequest);
 	}
-	
+
 	private void updateData(Cliente obj, Cliente cli) {
 		obj.setCpfOuCnpj(cli.getCpfOuCnpj());
 		obj.setTipo(cli.getTipo());
